+++ packages/react-table/src/docs/demos/table-demos/SortableResponsive.jsx
@@ -7,10 +7,19 @@ import {
   DropdownList,
   Flex,
   FlexItem,
+  Label,
   MenuToggle,
   MenuToggleElement,
+  OverflowMenu,
+  OverflowMenuContent,
+  OverflowMenuControl,
+  OverflowMenuDropdownItem,
+  OverflowMenuGroup,
+  OverflowMenuItem,
   PageSection,
+  PageSectionVariants,
   Pagination,
+  Select,
   SelectOption,
   SelectList,
   SelectGroup,
@@ -31,8 +40,7 @@ import SortAmountDownIcon from '@patternfly/react-icons/dist/esm/icons/sort-amou
 import CubeIcon from '@patternfly/react-icons/dist/esm/icons/cube-icon';
 import DashboardWrapper from '@patternfly/react-core/src/demos/examples/DashboardWrapper';
 import EllipsisVIcon from '@patternfly/react-icons/dist/esm/icons/ellipsis-v-icon';
-
-import { rows, columns } from './sampleData';
+import { rows, columns } from '@patternfly/react-table/src/docs/demos/table-demos/sampleData';
 
 export const ComposableTableSortable = () => {
   const [isKebabDropdownOpen, setIsKebabDropdownOpen] = React.useState(false);
@@ -89,7 +97,7 @@ export const ComposableTableSortable = () => {
     setPerPage(newPerPage);
   };
 
-  renderPagination = (variant) => (
+  const renderPagination = (variant) => (
     <Pagination
       isCompact
       itemCount={rows.length}